+++ src/main/java/com/cos/photogramstart/domain/likes/repository/PostLIkeRepository.java
@@ -9,11 +9,11 @@ public interface PostLIkeRepository extends JpaRepository<PostLike, Integer> {
 
     @Modifying
     @Query(value = "insert into PostLikes(imageId , userId ,createDate) values(:imageId , :principalId , now())" , nativeQuery = true)
-    int mLike(int imageId , int principalId);
+    int mLike(int imageId , Long principalId);
 
     @Modifying
     @Query(value = "delete from PostLikes where imageId =:imageId and userId =:principalId" , nativeQuery = true)
-    int mUnLike(int imageId , int principalId);
+    int mUnLike(int imageId , long principalId);
 
 //    PostLike findByImageIdAndUserId(int imageId , int userId);
 }
