+++ src/main/java/com/cos/photogramstart/web/dto/user/UserProfileUpdateResponse.java
@@ -1,5 +1,6 @@
 package com.cos.photogramstart.web.dto.user;
 
+import com.querydsl.core.annotations.QueryProjection;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -10,11 +11,13 @@ import lombok.NoArgsConstructor;
 @Data
 @Builder
 public class UserProfileUpdateResponse {
-    private int id;
+    private long id;
     private String username;
     private String name;
     private String bio;
     private String email;
     private String website;
     private String profileImageUrl;
+
+
 }
